requirements.txt
backports.cached-property==1.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
feedly-client @ file:///C:/Users/PC/Documents/ProgrammingTutorials/news%20aggregator/python-api-client
feedparser==6.0.8
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
protobuf==3.19.4
PyMySQL==1.0.2
python-dotenv==0.19.2
requests==2.27.1
sgmllib3k==1.0.0
SQLAlchemy==1.4.31
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
